Now it works with optimization on
The previous version of camd.h did not work properly with variadic parameter lists (tag lists).
In best case, it would only pass the first tag to the function (CreateMidi used to crash or hang).

Now with the macro based version, it'll build a temporary local  array from the macro parameters
and pass that along into the CreateMidiA function:

	LP1(0x2a, struct MidiNode *, CreateMidiA, struct TagItem *, tags, a0, \
	, CAMD_BASE_NAME)

	({ULONG _tags[] = {tags}; CreateMidiA((struct TagItem *) _tags);})